various improvements
use singleton for pages to boost load speed and save status, and use initialized instead of loaded for pages to avoid repeated loading procedures;
add an interface to pages and make the bottom buttons functional for different pages;
move bottom command bar from scooplist to mainwindow and make the buttons functional for each page (auto hide when switching to settings page);
there is now only one progress ring on the main window;
add load animation for settings page.
bugfix:
the theme toggle in settings page is not turned on when the theme is already dark.